requirements.txt
pymupdf>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
reportlab>=4.0.0
language-tool-python>=2.7.3
pyspellchecker>=0.7.2
